Se termino la funcionabilidad de las familias de productos

+++ android/app/src/main/java/com/marcosilv7/narutodelivery/dto/ProductFamilyDTO.java
@@ -4,7 +4,7 @@ public class ProductFamilyDTO {
     private Long id;
     private String name;
     private String image;
-    private Integer countProducts;
+    private Long countProducts;
 
     public ProductFamilyDTO(){}
 
@@ -32,11 +32,11 @@ public class ProductFamilyDTO {
         this.image = image;
     }
 
-    public Integer getCountProducts() {
+    public Long getCountProducts() {
         return countProducts;
     }
 
-    public void setCountProducts(Integer countProducts) {
+    public void setCountProducts(Long countProducts) {
         this.countProducts = countProducts;
     }
 }
